refactor: extract prediction composables and use $fetch
Move the form/validation and prediction request logic out of the
PredictionPage view component into focused composables
(usePredictionForm, usePrediction), keeping the view thin per Vue/Nuxt
conventions. Replace the native fetch call and its hand-rolled response
parser with Nuxt's $fetch, which throws a typed error and removes the
bespoke error-extraction code.
